View injects layout config and component config
